app_backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.3
watchfiles